Add the actual key algorithm that works for everything outside of DDR Solo Bass Mix (seemingly)
