+++ py573a/py573a_native.py
@@ -28,13 +28,19 @@ def get_key_information(sha1):
     song = db.get(sha1, None)
 
     if song:
-        return (bytearray(base64.b64decode(song['key'])), bytearray(base64.b64decode(song['scramble'])), song['counter'])
+        if 'key' in song:
+            return (bytearray(base64.b64decode(song['key'])), None, None)
+
+        else:
+            return (song['key1'], song['key2'], song['key3'])
 
     return (None, None, None)
 
 
-def decrypt(data, key, scramble, counter):
+def decrypt_ddrsbm(data, key):
     output_data = bytearray(len(data))
+
+    scramble = bytearray([key[-1]]) + key[:-1]
 
     data_len = len(data) // 2
     key_len = len(key)
@@ -53,81 +59,96 @@ def decrypt(data, key, scramble, counter):
             is_odd_bit_set = int((cur_data & (1 << odd_bit_shift)) != 0)
             is_key_bit_set = int((key[idx % key_len] & (1 << cur_bit)) != 0)
             is_scramble_bit_set = int((scramble[idx % scramble_len] & (1 << cur_bit)) != 0)
-            is_counter_bit_set = int((counter & (1 << cur_bit)) != 0)
-            is_counter_bit_inv_set =  int(counter & (1 << ((7 - cur_bit) & 0xff)) != 0)
 
             if is_scramble_bit_set == 1:
                 is_even_bit_set, is_odd_bit_set = is_odd_bit_set, is_even_bit_set
 
-            if ((is_even_bit_set ^ is_counter_bit_inv_set ^ is_key_bit_set)) == 1:
+            if ((is_even_bit_set ^ is_key_bit_set)) == 1:
                 output_word |= 1 << even_bit_shift
 
-            if (is_odd_bit_set ^ is_counter_bit_set) == 1:
+            if is_odd_bit_set == 1:
                 output_word |= 1 << odd_bit_shift
 
         output_data[output_idx] = (output_word >> 8) & 0xff
         output_data[output_idx+1] = output_word & 0xff
         output_idx += 2
 
-        counter = (counter + 1) & 0xff
-
     return bytearray(output_data)
 
 
-def encrypt(data, key, scramble, counter):
+# You crazy for this one
+# Thanks anon and RC
+def decrypt(data, key1, key2, key3):
+    def is_bit_set(value, n):
+        return (value >> n) & 1
+
+    def bit_swap(v, b15, b14, b13, b12, b11, b10, b9, b8, b7, b6, b5, b4, b3, b2, b1, b0):
+        return (is_bit_set(v, b15) << 15) | (is_bit_set(v, b14) << 14) | (is_bit_set(v, b13) << 13) | (is_bit_set(v, b12) << 12) |\
+               (is_bit_set(v, b11) << 11) | (is_bit_set(v, b10) << 10) | (is_bit_set(v, b9) << 9)   | (is_bit_set(v, b8) << 8)   |\
+               (is_bit_set(v, b7) << 7)   | (is_bit_set(v, b6) << 6)   | (is_bit_set(v, b5) << 5)   | (is_bit_set(v, b4) << 4)   |\
+               (is_bit_set(v, b3) << 3)   | (is_bit_set(v, b2) << 2)   | (is_bit_set(v, b1) << 1)   | (is_bit_set(v, b0) << 0)
+
     output_data = bytearray(len(data))
 
     data_len = len(data) // 2
-    key_len = len(key)
-    scramble_len = len(scramble)
-    output_idx = 0
 
     for idx in range(0, data_len):
-        output_word = 0
-        cur_data = (data[(idx * 2)] << 8) | data[(idx * 2) + 1]
+        v = (data[idx * 2 + 1] << 8) | data[idx * 2]
 
-        for cur_bit in range(0, 8):
-            even_bit_shift = (cur_bit * 2) & 0xff
-            odd_bit_shift = (cur_bit * 2 + 1) & 0xff
+        m = key1 ^ key2
 
-            is_even_bit_set = int((cur_data & (1 << even_bit_shift)) != 0)
-            is_odd_bit_set = int((cur_data & (1 << odd_bit_shift)) != 0)
-            is_key_bit_set = int((key[idx % key_len] & (1 << cur_bit)) != 0)
-            is_scramble_bit_set = int((scramble[idx % scramble_len] & (1 << cur_bit)) != 0)
-            is_counter_bit_set = int((counter & (1 << cur_bit)) != 0)
-            is_counter_bit_inv_set =  int(counter & (1 << ((7 - cur_bit) & 0xff)) != 0)
+        v = bit_swap(
+            v,
+            15 - is_bit_set(m, 0xF),
+            14 + is_bit_set(m, 0xF),
+            13 - is_bit_set(m, 0xE),
+            12 + is_bit_set(m, 0xE),
+            11 - is_bit_set(m, 0xB),
+            10 + is_bit_set(m, 0xB),
+            9 - is_bit_set(m, 0x9),
+            8 + is_bit_set(m, 0x9),
+            7 - is_bit_set(m, 0x8),
+            6 + is_bit_set(m, 0x8),
+            5 - is_bit_set(m, 0x5),
+            4 + is_bit_set(m, 0x5),
+            3 - is_bit_set(m, 0x3),
+            2 + is_bit_set(m, 0x3),
+            1 - is_bit_set(m, 0x2),
+            0 + is_bit_set(m, 0x2)
+        )
 
-            set_even_bit = is_even_bit_set ^ is_counter_bit_inv_set ^ is_key_bit_set
-            set_odd_bit = is_odd_bit_set ^ is_counter_bit_set
+        for p in [(0x0d, 14), (0x0c, 12), (0x0a, 10), (0x07, 8), (0x06, 6), (0x04, 4), (0x01, 2), (0x00, 0)]:
+            v ^= is_bit_set(m, p[0]) << p[1]
 
-            if is_scramble_bit_set == 1:
-                set_even_bit, set_odd_bit = set_odd_bit, set_even_bit
+        v &= 0xffff
 
-            if set_even_bit == 1:
-                output_word |= 1 << even_bit_shift
+        v ^= bit_swap(
+                key3,
+                7, 0, 6, 1,
+                5, 2, 4, 3,
+                3, 4, 2, 5,
+                1, 6, 0, 7
+            )
 
-            if set_odd_bit == 1:
-                output_word |= 1 << odd_bit_shift
+        output_data[idx * 2] = (v >> 8) & 0xff
+        output_data[idx * 2 + 1] = v & 0xff
 
-        output_data[output_idx] = output_word & 0xff
-        output_data[output_idx+1] = (output_word >> 8) & 0xff
-        output_idx += 2
+        key1 = ((key1 & 0x8000) | ((key1 << 1) & 0x7FFE) | ((key1 >> 14) & 1)) & 0xFFFF
 
-        counter = (counter + 1) & 0xff
+        if (((key1 >> 15) ^ key1) & 1) != 0:
+            key2 = ((key2 << 1) | (key2 >> 15)) & 0xFFFF
+
+        key3 += 1
 
     return bytearray(output_data)
 
 
-if __name__ == "__main__":
+def main():
     parser = argparse.ArgumentParser()
 
     parser.add_argument('--input', help='Input file', default=None)
     parser.add_argument('--output', help='Output file', default=None)
     parser.add_argument('--sha1', help='Force usage of a specific SHA-1 for encryption keys (optional)', default=None)
-
-    group = parser.add_mutually_exclusive_group(required=True)
-    group.add_argument('--decrypt', help='Decrypt input file', action='store_true')
-    group.add_argument('--encrypt', help='Encrypt input file', action='store_true')
 
     args = parser.parse_args()
 
@@ -149,7 +170,7 @@ if __name__ == "__main__":
         data = infile.read()
 
     sha1 = args.sha1
-    if args.decrypt and sha1 is None:
+    if sha1 is None:
         m = hashlib.sha1()
         m.update(data)
         sha1 = m.hexdigest()
@@ -159,15 +180,19 @@ if __name__ == "__main__":
 
     print("Using SHA-1:", sha1)
 
-    key, scramble, counter = get_key_information(sha1)
-    if key is None or scramble is None or counter is None:
+    key1, key2, key3 = get_key_information(sha1)
+    if key1 is None:
         raise Exception("Couldn't find key information for file with SHA-1 hash of %s" % (sha1))
 
-    if args.decrypt:
-        output_data = decrypt(data, key, scramble, counter)
+    if isinstance(key1, int) and isinstance(key2, int) and isinstance(key3, int):
+        output_data = decrypt(data, key1, key2, key3)
 
-    elif args.encrypt:
-        output_data = encrypt(data, key, scramble, counter)
+    else:
+        output_data = decrypt_ddrsbm(data, key1)
 
     with open(args.output, "wb") as outfile:
         outfile.write(output_data)
+
+
+if __name__ == "__main__":
+    main()
