+++ py573a/setup.py
@@ -2,5 +2,5 @@ from distutils.core import setup
 from Cython.Build import cythonize
 
 setup(
-    ext_modules = cythonize(["enc573.pyx", "bruteforce.pyx"], annotate=True, language_level=3)
+    ext_modules = cythonize(["enc573.pyx"], annotate=True, language_level=3)
 )
