+++ py573a_old/setup.py
@@ -1,0 +1,6 @@
+from distutils.core import setup
+from Cython.Build import cythonize
+
+setup(
+    ext_modules = cythonize(["enc573.pyx", "bruteforce.pyx"], annotate=True, language_level=3)
+)
